Lab Solution
Comments on QUESTION 10:

 I copied the solution from another folder, but understood what is
doing. Very hard.  My rationale of what is doing is:
 - Starts the loop iterating from 0 and looking for that value in the
dictionary, if it finds it, it unwraps the value and execute the search
for high score.  If it doesn't it assigns the next integer (1) to grade
and searches again.
 - It able to unwrap the value of the lowest grade in this case 9, it
compares the highGrade string to the dictionary pair, in the case of 9
is compared "B" to "B" and asssignes itself the highGrade variable.
 - IMPORTANT:  Strings can be compared in boolean terms like numbers,
with Z being the largest and A the lowest. See examples above
 - That is the reason with the "if score < highScore {" assigned the
less than statement to the highest grade.
 - Finally the code prints out the statement but only if is able to
unwrap the highGrade value, protecting the print against a nil.
 - The code cannot deal with ties, if the comparator between "score and
highScore" is "<" then it will output the highGrade to the lowest grade
obtaining "A" or whatever the highest score is, if the comparator is
"<=" it will output the highest grade obtaining the highest score.